+++ experimental/packages/opentelemetry-instrumentation-grpc/src/utils.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { SpanStatusCode, SpanStatus } from '@opentelemetry/api';
+import {SpanStatusCode, SpanStatus, Span} from '@opentelemetry/api';
 import type * as grpcTypes from 'grpc';
 import type * as grpcJsTypes from '@grpc/grpc-js';
 import { IgnoreMatcher } from './types';
@@ -110,3 +110,28 @@ export const _extractMethodAndService = (name: string): { service: string, metho
     method
   });
 };
+
+
+export function metadataCapture(type: 'request' | 'response', metadataToAdd: string[]) {
+  const normalizedMetadataAttributes = new Map(metadataToAdd.map(value => [value.toLowerCase(), value.toLowerCase().replace(/-/g, '_')]));
+
+  return (span: Span, metadata: grpcJsTypes.Metadata | grpcTypes.Metadata) => {
+    for (const [capturedMetadata, normalizedMetadata] of normalizedMetadataAttributes) {
+      const metadataValues =
+        metadata
+          .get(capturedMetadata)
+          .flatMap(value =>
+            typeof value === 'string' ? value.toString() : []
+          );
+
+      if (metadataValues === undefined || metadataValues === []) {
+        continue;
+      }
+
+      const key = `rpc.${type}.metadata.${normalizedMetadata}`;
+
+      span.setAttribute(key, metadataValues);
+    }
+  };
+}
+